+++ kyu_6/default_list/test_default_list.py
@@ -297,47 +297,3 @@ class DefaultListTestCase(unittest.TestCase):
             actual = lst[i]
             print_log(lst=lst, i=i, expected=expected, actual=actual)
             self.assertEqual(expected, actual)
-
-    def test_default_list_edge_cases_pop(self):
-        """
-        Testing 'DefaultList' class: pop.
-
-        Tests regular pop operations and edge case of popping from empty list.
-        :return:
-        """
-        # pylint: disable-msg=R0801
-        allure.dynamic.title("Testing 'DefaultList' class: edge cases for pop")
-        allure.dynamic.severity(allure.severity_level.NORMAL)
-        allure.dynamic.description_html(
-            '<h3>Codewars badge:</h3>'
-            '<img src="https://www.codewars.com/users/myFirstCode'
-            '/badges/large">'
-            '<h3>Test Description:</h3>'
-            "<p>Testing edge cases for pop method.</p>")
-        # pylint: enable-msg=R0801
-        with allure.step("Create a default/empty list:"):
-            default_value = 'default'
-            empty_list = DefaultList([], default_value)
-
-        with allure.step("Test popping from empty list without index:"):
-            actual = empty_list.pop()
-            print_log(lst=empty_list, i=None, expected=default_value, actual=actual)
-            self.assertEqual(default_value, actual)
-
-        with allure.step("Test popping from empty list with index:"):
-            actual = empty_list.pop(0)
-            print_log(lst=empty_list, i=0, expected=default_value, actual=actual)
-            self.assertEqual(default_value, actual)
-
-        with allure.step("Test popping from list with elements:"):
-            test_list = DefaultList([1, 2, 3], default_value)
-
-        with allure.step("Test popping from list with elements and no index:"):
-            actual = test_list.pop()
-            print_log(lst=test_list, i=None, expected=3, actual=actual)
-            self.assertEqual(3, actual)
-
-        with allure.step("Test popping from list with elements and index:"):
-            actual = test_list.pop(0)
-            print_log(lst=test_list, i=0, expected=1, actual=actual)
-            self.assertEqual(1, actual)
